refactor: changed some items in the table component of the courses page
- Renamed the file to "coursesTable".
- Removed "Next" and "Previous" buttons.
- Added a scroll bar on the side of the Table body.
